// frontend/src/utils/common-utils.d.ts
import { OperationName, Page } from "../config/config";
export declare class CommonUtils {
    static showTitlePage(page: Page, type: OperationName): void;
    static showCategories(typeElement: HTMLSelectElement, categoryElement: HTMLSelectElement): Promise<void>;
    static updateModal(type: OperationName, url: string): void;
}
//# sourceMappingURL=common-utils.d.ts.map